# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Moxxy Agentic Framework API
  version: 0.1.0
  description: Local-first agentic framework REST API
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "mox_*"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
paths:
...
  /v1/hooks/{token}:
    post:
      tags: [Webhooks]
      summary: Receive inbound webhook
      description: Unauthenticated endpoint. Uses HMAC-SHA256 signature verification via x-hub-signature-256 header.
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: x-hub-signature-256
          in: header
          required: true
          schema:
            type: string
        - name: x-github-event
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Webhook accepted or filtered
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [accepted, filtered]
                  run_id:
                    type: string
                    nullable: true
                  delivery_id:
                    type: string
                  message:
                    type: string
        '401':
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Unknown webhook token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: Webhook disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/agents/{id}/webhooks:
    get:
      tags: [Webhooks]
      summary: List webhooks for agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    agent_id:
                      type: string
                    label:
                      type: string
                    url:
                      type: string
                    event_filter:
                      type: string
                      nullable: true
                      description: Comma-separated event types
                    enabled:
                      type: boolean
                    created_at:
                      type: string
                    updated_at:
                      type: string
  /v1/agents/{id}/webhooks/{wh_id}:
    delete:
      tags: [Webhooks]
      summary: Delete webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: wh_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [deleted]
                  id:
                    type: string
        '404':
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...